pyproject.toml
[project]
name = "GeminiChatbot"
version = "0.1.0"
description = "This is a FastAPI-based web application that provides a chat interface powered by Google's Gemini AI model through LangChain."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-google-genai>=2.0.10",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
